requirements.txt
cloudpickle==2.2.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
gym==0.26.2
gym-notices==0.0.8
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tensordict==0.0.1rc0
torch==1.13.1
torchrl==0.0.3
typing_extensions==4.4.0
cloudpickle==2.2.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
gym==0.26.2
gym-notices==0.0.8
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
Pillow==9.4.0
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tensordict==0.0.1rc0
torch==1.13.1
torchrl==0.0.3
typing_extensions==4.4.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudpickle==2.2.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
gym==0.26.2
gym-notices==0.0.8
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
Pillow==9.4.0
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
tensordict==0.0.1rc0
torch==1.13.1
torchrl==0.0.3
torchvision==0.14.1
typing_extensions==4.4.0
urllib3==1.26.14
